+++ src/ui/MessageContent/MessageContentForTemplateMessage.tsx
